working MELD unittests
