+++ tests/coupled/Makefile
@@ -1,0 +1,9 @@
+include ../../Makefile.in
+
+default: meld_ucrm
+
+clean:
+	rm *.out || echo "no *.out files to remove"
+
+%: %.cu
+	${GPU_CXX} ${GPU_CC_FLAGS} -DUSE_GPU -Xcompiler ${CC_FLAGS} -DUSE_CUSPARSE ${INCLUDE_FLAGS} -std=c++17 $< -o $@.out ${LINK}
