--- tests/coupled/meld_plate/meld_cpu.py
@@ -1,97 +1,0 @@
-import numpy as np
-from funtofem import TransferScheme
-from mpi4py import MPI
-import matplotlib.pyplot as plt
-
-np.random.seed(1234567)
-
-def make_grid_mesh(nnx, nny, Lx, Ly, z0):
-    N = nnx * nny
-    dx = Lx / (nnx - 1)
-    dy = Ly / (nny - 1)
-    x0 = np.zeros((3 * N,))
-
-    for iy in range(nny):
-        yfac = np.sin(np.pi * iy / (nny - 1))
-        for ix in range(nnx):
-            xfac = np.sin(np.pi * ix / (nnx - 1))
-            ind = iy * nnx + ix
-            x0[3 * ind] = ix * dx
-            x0[3*ind+1] = iy * dy
-            x0[3*ind+2] = z0 * xfac * yfac
-    
-    return x0
-
-def make_inplane_shear_disp(x0, angleDeg):
-    N = x0.shape[0] // 3
-    u = np.zeros((3 * N,))
-    angleRad = angleDeg * np.pi / 180.0
-
-    for inode in range(N):
-        u[3*inode] = np.tan(angleRad / 2.0) * x0[3*inode+1]
-        u[3*inode+1] = np.tan(angleRad / 2.0) * x0[3*inode]
-        u[3*inode+2] = 0.0
-    return u
-
-def plot_quantity(xpts, vars, title:str):
-    # assume xpts, vars are same size and only plot xy plane behavior
-    # with subplots
-    fig, ax = plt.subplots(1, 2, figsize=(20,10))
-    N = xpts.shape[0] // 3
-    n = int(np.sqrt(N))
-    X = xpts[0::3].reshape((n,n))
-    Y = xpts[1::3].reshape((n,n))
-    plt.title(title)
-    for i in range(2):
-        vd = vars[i::3].reshape((n,n))
-        # 'RdBu_r'
-        ax[i].contourf(X, Y, vd, antialiased=True, cmap='coolwarm', levels=100)
-    # plt.show()
-    plt.savefig(f"out/{title}.png", dpi=400)
-
-
-if __name__=="__main__":
-    Lx = Ly = 1.0
-    nna = 30
-    nns = 17
-    xs0 = make_grid_mesh(nns, nns, Lx, Ly, 0.01)
-    xa0 = make_grid_mesh(nna, nna, Lx, Ly, 0.01)
-    na = xa0.shape[0] // 3; ns = xs0.shape[0] // 3
-
-    us = make_inplane_shear_disp(xs0, 20.0)
-    fa = make_inplane_shear_disp(xa0, 10.0)
-    ua = xa0 * 0.0
-    fs = us * 0.0
-
-    # make MELD object
-    comm = MPI.COMM_WORLD
-    isymm = -1; nn = 32; beta = 10.0
-    meld = TransferScheme.pyMELD(comm, comm, 0, comm, 0, isymm, nn, beta)
-
-    meld.setAeroNodes(xa0)
-    meld.setStructNodes(xs0)
-    meld.initialize()
-
-    # set displacements
-    meld.transferDisps(us, ua)
-    meld.transferLoads(fa, fs)
-
-    np.set_printoptions(threshold=np.inf)
-    print(f"{us[:10]=}")
-    # print(f"{ua[:2700]=}")
-
-    ua_gpu = np.array([4.69456e-03,4.69456e-03,2.03921e-05,5.01792e-03,7.35143e-03,1.27455e-05,4.88445e-03,1.23758e-02,5.31521e-06,4.42136e-03,1.82893e-02,2.88366e-07,4.03132e-03,2.43504e-02,-2.95306e-06,3.81639e-03,3.04401e-02,-5.20311e-06,3.70125e-03,3.65044e-02,-7.59291e-06,3.71026e-03,4.25989e-02,-9.34249e-06,3.69192e-03,4.86687e-02,-1.14864e-05,3.71215e-03,5.47547e-02,-1.31721e-05,3.68945e-03,6.08350e-02,-1.47458e-05,3.70983e-03,6.69100e-02,-1.64310e-05,3.69328e-03,7.30007e-02,-1.73150e-05,3.70696e-03,7.90669e-02,-1.88929e-05,3.70125e-03,8.51635e-02,-1.91740e-05,3.70093e-03,9.12267e-02,-2.03516e-05,3.70721e-03,9.73228e-02,-2.02586e-05,3.69302e-03,1.03390e-01,-2.07403e-05,3.70991e-03,1.09479e-01,-2.04692e-05,3.68936e-03,1.15556e-01,-2.01085e-05,3.71203e-03,1.21634e-01,-1.97350e-05,3.69203e-03,1.27722e-01,-1.86050e-05,3.70997e-03,1.33790e-01,-1.79894e-05,3.70148e-03,1.39885e-01,-1.63683e-05,3.81608e-03,1.45951e-01,-1.58601e-05,4.03141e-03,1.52044e-01,-1.51352e-05,4.42089e-03,1.58112e-01,-1.53721e-05,4.87923e-03,1.64022e-01,-1.69262e-05,5.00525e-03,1.69008e-01,-2.12157e-05,4.69457e-03,1.71632e-01,-2.91226e-05,7.35143e-03,5.01792e-03,1.27455e-05,7.23911e-03,7.23911e-03,7.73485e-06,6.99803e-03,1.22734e-02,3.09525e-06,6.76317e-03,1.82560e-02,1.38884e-06,6.58378e-03,2.42919e-02,1.45502e-07,6.51326e-03,3.04434e-02,7.89539e-07,6.49899e-03,3.64387e-02,-9.96209e-07,6.50057e-03,4.25956e-02,2.32940e-07,6.50214e-03,4.86184e-02,-1.35922e-06,6.50258e-03,5.47315e-02,-8.44323e-07,6.50509e-03,6.08072e-02,-1.23685e-06,6.50275e-03,6.68641e-02,-2.17144e-06,6.50067e-03,7.29943e-02,-8.96096e-07,6.49898e-03,7.90033e-02,-3.31965e-06,6.49814e-03,8.51687e-02,-8.15207e-07,6.49811e-03,9.11586e-02,-3.85318e-06,6.49900e-03,9.73240e-02,-1.29651e-06,6.50063e-03,1.03333e-01,-3.64564e-06,6.50275e-03,1.09463e-01,-2.21255e-06,6.50507e-03,1.15520e-01,-2.96155e-06,6.50257e-03,1.21596e-01,-3.07651e-06,6.50215e-03,1.27709e-01,-2.21509e-06,6.50054e-03,1.33732e-01,-3.50349e-06,6.49901e-03,1.39889e-01,-1.79228e-06,6.51322e-03,1.45884e-01,-3.33803e-06,6.58380e-03,1.52036e-01,-2.47009e-06,6.76313e-03,1.58072e-01,-3.83536e-06,6.99782e-03,1.64055e-01,-5.45667e-06,7.23909e-03,1.69088e-01,-1.10463e-05,7.31861e-03,1.71322e-01,-2.12157e-05,1.23758e-02,4.88445e-03,5.31520e-06,1.22734e-02,6.99803e-03,3.09526e-06,1.22033e-02,1.22033e-02,4.73745e-07,1.21551e-02,1.82831e-02,3.48701e-07,1.21242e-02,2.42411e-02,-1.09776e-06,1.21005e-02,3.05113e-02,1.47996e-06,1.21013e-02,3.63815e-02,-1.91190e-06,1.21003e-02,4.26484e-02,1.71611e-06,1.21017e-02,4.85878e-02,-1.25754e-06,1.21011e-02,5.47467e-02,6.58799e-07,1.21028e-02,6.08133e-02,3.88088e-07,1.21009e-02,6.68342e-02,-1.22828e-06,1.21015e-02,7.30352e-02,2.35477e-06,1.21004e-02,7.89432e-02,-2.79778e-06,1.21011e-02,8.52305e-02,3.63802e-06,1.21011e-02,9.10964e-02,-2.99703e-06,1.21004e-02,9.73837e-02,3.48465e-06,1.21014e-02,1.03292e-01,-1.77571e-06,1.21009e-02,1.09493e-01,1.88088e-06,1.21028e-02,1.15514e-01,1.59408e-07,1.21011e-02,1.21580e-01,-1.51042e-07,1.21017e-02,1.27739e-01,1.83348e-06,1.21003e-02,1.33679e-01,-1.41284e-06,1.21013e-02,1.39945e-01,2.49556e-06,1.21005e-02,1.45816e-01,-1.36773e-06,1.21242e-02,1.52086e-01,1.43968e-06,1.21551e-02,1.58044e-01,-4.27264e-07,1.22033e-02,1.64124e-01,-6.76569e-07,1.22721e-02,1.69329e-01,-5.45667e-06,1.23051e-02,1.71448e-01,-1.69262e-05,1.82893e-02,4.42136e-03,2.88367e-07,1.82560e-02,6.76317e-03,1.38884e-06,1.82831e-02,1.21551e-02,3.48701e-07,1.83339e-02,1.83339e-02,1.49981e-06,1.83743e-02,2.41540e-02,-9.11496e-07,1.83870e-02,3.05875e-02,3.55330e-06,1.83890e-02,3.63123e-02,-2.14424e-06,1.83826e-02,4.27074e-02,3.42096e-06,1.83905e-02,4.85512e-02,-1.62243e-06,1.83796e-02,5.47636e-02,1.19234e-06,1.83909e-02,6.08210e-02,4.99579e-07,1.83799e-02,6.68020e-02,-2.40852e-06,1.83906e-02,7.30845e-02,3.17017e-06,1.83847e-02,7.88751e-02,-5.56127e-06,1.83889e-02,8.53051e-02,4.82633e-06,1.83889e-02,9.10219e-02,-6.46499e-06,1.83848e-02,9.74519e-02,4.12376e-06,1.83906e-02,1.03242e-01,-4.84556e-06,1.83800e-02,1.09525e-01,1.12804e-06,1.83909e-02,1.15506e-01,-1.95772e-06,1.83795e-02,1.21564e-01,-2.51001e-06,1.83905e-02,1.27776e-01,6.08908e-07,1.83824e-02,1.33620e-01,-4.85905e-06,1.83890e-02,1.40015e-01,1.66788e-06,1.83869e-02,1.45739e-01,-4.95456e-06,1.83743e-02,1.52173e-01,8.59377e-07,1.83338e-02,1.57993e-01,-2.14183e-06,1.82831e-02,1.64172e-01,-4.27268e-07,1.82550e-02,1.69564e-01,-3.83536e-06,1.82148e-02,1.71906e-01,-1.53721e-05,2.43504e-02,4.03132e-03,-2.95306e-06,2.42919e-02,6.58378e-03,1.45500e-07,2.42411e-02,1.21242e-02,-1.09777e-06,2.41540e-02,1.83743e-02,-9.11482e-07,2.41055e-02,2.41055e-02,-4.27220e-06,2.40991e-02,3.06252e-02,9.23435e-07,2.40923e-02,3.62769e-02,-4.88565e-06,2.41078e-02,4.27381e-02,1.87051e-06,2.40828e-02,4.85300e-02,-3.41169e-06,2.41141e-02,5.47724e-02,4.51318e-07,2.40796e-02,6.08256e-02,1.33760e-07,2.41137e-02,6.67850e-02,-2.56220e-06,2.40862e-02,7.31119e-02,4.33383e-06,2.41049e-02,7.88404e-02,-5.04787e-06,2.40965e-02,8.53416e-02,7.19725e-06,2.40964e-02,9.09852e-02,-4.98574e-06,2.41050e-02,9.74865e-02,7.35830e-06,2.40861e-02,1.03215e-01,-2.22623e-06,2.41137e-02,1.09542e-01,4.68837e-06,2.40796e-02,1.15501e-01,1.97993e-06,2.41141e-02,1.21554e-01,1.15941e-06,2.40830e-02,1.27797e-01,5.58831e-06,2.41078e-02,1.33589e-01,-9.82075e-07,2.40925e-02,1.40050e-01,7.04433e-06,2.40991e-02,1.45702e-01,-6.63223e-07,2.41056e-02,1.52221e-01,6.10016e-06,2.41540e-02,1.57953e-01,8.59384e-07,2.42411e-02,1.64203e-01,1.43967e-06,2.42914e-02,1.69743e-01,-2.47010e-06,2.42825e-02,1.72296e-01,-1.51352e-05,3.04401e-02,3.81639e-03,-5.20311e-06,3.04434e-02,6.51326e-03,7.89540e-07,3.05113e-02,1.21005e-02,1.47996e-06,3.05875e-02,1.83870e-02,3.55331e-06,3.06252e-02,2.40991e-02,9.23435e-07,3.06291e-02,3.06291e-02,5.41247e-06,3.06374e-02,3.62725e-02,-6.09341e-07,3.06204e-02,4.27412e-02,4.99476e-06,3.06498e-02,4.85268e-02,-4.83266e-07,3.06150e-02,5.47735e-02,2.23543e-06,3.06536e-02,6.08265e-02,1.44342e-06,3.06157e-02,6.67831e-02,-2.14388e-06,3.06470e-02,7.31165e-02,3.96731e-06,3.06244e-02,7.88369e-02,-6.09727e-06,3.06341e-02,8.53463e-02,5.27298e-06,3.06341e-02,9.09807e-02,-7.64170e-06,3.06244e-02,9.74901e-02,4.12044e-06,3.06469e-02,1.03211e-01,-6.53390e-06,3.06158e-02,1.09544e-01,4.08155e-07,3.06536e-02,1.15501e-01,-3.68705e-06,3.06149e-02,1.21554e-01,-4.13488e-06,3.06499e-02,1.27800e-01,-9.31080e-07,3.06204e-02,1.33586e-01,-7.24787e-06,3.06374e-02,1.40054e-01,1.83952e-07,3.06290e-02,1.45698e-01,-7.72852e-06,3.06252e-02,1.52228e-01,-6.63222e-07,3.05875e-02,1.57940e-01,-4.95457e-06,3.05113e-02,1.64227e-01,-1.36770e-06,3.04431e-02,1.69814e-01,-3.33802e-06,3.03759e-02,1.72511e-01,-1.58601e-05,3.65044e-02,3.70125e-03,-7.59291e-06,3.64387e-02,6.49899e-03,-9.96206e-07,3.63815e-02,1.21013e-02,-1.91189e-06,3.63123e-02,1.83890e-02,-2.14423e-06,3.62769e-02,2.40923e-02,-4.88566e-06,3.62725e-02,3.06374e-02,-6.09228e-07,3.62620e-02,3.62620e-02,-5.50211e-06,3.62825e-02,4.27462e-02,5.43485e-07,3.62499e-02,4.85206e-02,-4.15690e-06,3.62868e-02,5.47753e-02,-3.92725e-07,3.62466e-02,6.08279e-02,-6.85311e-07,3.62863e-02,6.67797e-02,-2.85173e-06,3.62532e-02,7.31250e-02,3.54885e-06,3.62798e-02,7.88312e-02,-4.86486e-06,3.62677e-02,8.53561e-02,6.48173e-06,3.62677e-02,9.09707e-02,-4.69041e-06,3.62799e-02,9.74955e-02,6.84543e-06,3.62532e-02,1.03202e-01,-1.89536e-06,3.62865e-02,1.09547e-01,4.71396e-06,3.62466e-02,1.15499e-01,2.45012e-06,3.62868e-02,1.21551e-01,1.62056e-06,3.62500e-02,1.27806e-01,6.25411e-06,3.62824e-02,1.33581e-01,-2.43102e-07,2.33671e+00,3.82326e+00,2.48364e-02,3.62725e-02,1.45690e-01,1.83956e-07,3.62770e-02,1.52235e-01,7.04441e-06,3.63123e-02,1.57938e-01,1.66790e-06,3.63816e-02,1.64226e-01,2.49552e-06,3.64384e-02,1.69828e-01,-1.79229e-06,3.64417e-02,1.72626e-01,-1.63683e-05,4.25989e-02,3.71026e-03,-9.34249e-06,4.25956e-02,6.50057e-03,2.32923e-07,4.26484e-02,1.21003e-02,1.71615e-06,4.27074e-02,1.83826e-02,3.42096e-06,4.27381e-02,2.41078e-02,1.87064e-06,4.27412e-02,3.06204e-02,4.99476e-06,4.27462e-02,3.62825e-02,5.43354e-07,4.27320e-02,4.27320e-02,4.56380e-06,4.27502e-02,4.85353e-02,4.39189e-07,4.27268e-02,5.47707e-02,2.33078e-06,4.27532e-02,6.08250e-02,1.71931e-06,4.27283e-02,6.67886e-02,-1.23424e-06,4.27496e-02,7.31065e-02,3.51089e-06,4.27361e-02,7.88468e-02,-4.60403e-06,4.27450e-02,8.53371e-02,4.50071e-06,4.27449e-02,9.09899e-02,-6.13588e-06,4.27361e-02,9.74801e-02,3.39943e-06,4.27494e-02,1.03221e-01,-5.31942e-06,4.27283e-02,1.09538e-01,2.15375e-07,4.27531e-02,1.15502e-01,-3.19880e-06,4.27268e-02,1.21556e-01,-3.68023e-06,4.27504e-02,1.27792e-01,-1.09276e-06,4.27320e-02,1.33595e-01,-6.51684e-06,4.27464e-02,1.40045e-01,-2.43080e-07,4.27411e-02,1.45707e-01,-7.24785e-06,4.27382e-02,1.52219e-01,-9.82038e-07,4.27073e-02,1.57945e-01,-4.85904e-06,4.26485e-02,1.64227e-01,-1.41289e-06,4.25954e-02,1.69826e-01,-3.50349e-06,4.25369e-02,1.72617e-01,-1.79894e-05,4.86687e-02,3.69192e-03,-1.14864e-05,4.86184e-02,6.50214e-03,-1.35922e-06,4.85878e-02,1.21017e-02,-1.25757e-06,4.85512e-02,1.83905e-02,-1.62233e-06,4.85300e-02,2.40828e-02,-3.41191e-06,4.85268e-02,3.06498e-02,-4.83033e-07,4.85206e-02,3.62499e-02,-4.15690e-06,4.85353e-02,4.27502e-02,4.39454e-07,4.85139e-02,4.85139e-02,-3.30991e-06,4.85373e-02,5.47768e-02,-3.26424e-07,4.85121e-02,6.08296e-02,-6.39917e-07,4.85370e-02,6.67773e-02,-2.43058e-06,4.85159e-02,7.31344e-02,2.78398e-06,4.85332e-02,7.88252e-02,-4.29531e-06,4.85241e-02,8.53690e-02,5.20691e-06,4.85240e-02,9.09578e-02,-4.50154e-06,4.85335e-02,9.75016e-02,5.34504e-06,4.85159e-02,1.03193e-01,-2.31084e-06,4.85373e-02,1.09549e-01,3.45891e-06,4.85121e-02,1.15497e-01,1.36361e-06,4.85373e-02,1.21550e-01,6.87867e-07,1.58280e-01,2.37555e-01,4.28221e-06,4.85351e-02,1.33577e-01,-1.09268e-06,4.85208e-02,1.40077e-01,6.25441e-06,4.85266e-02,1.45677e-01,-9.31021e-07,4.85303e-02,1.52244e-01,5.58845e-06,4.85512e-02,1.57937e-01,6.08986e-07,4.85879e-02,1.64225e-01,1.83346e-06,4.86180e-02,1.69825e-01,-2.21511e-06,4.86054e-02,1.72635e-01,-1.86050e-05,5.47547e-02,3.71215e-03,-1.31721e-05,5.47315e-02,6.50258e-03,-8.44355e-07,5.47467e-02,1.21011e-02,6.58969e-07,5.47636e-02,1.83796e-02,1.19234e-06,5.47724e-02,2.41141e-02,4.51339e-07,5.47735e-02,3.06150e-02,2.23566e-06,5.47753e-02,3.62868e-02,-3.92496e-07,5.47707e-02,4.27268e-02,2.33090e-06,5.47768e-02,4.85373e-02,-3.26531e-07,5.47689e-02,5.47689e-02,1.14993e-06,5.47771e-02,6.08244e-02,8.32033e-07,5.47698e-02,6.67921e-02,-1.05207e-06,5.47773e-02,7.31044e-02,2.51985e-06,5.47723e-02,7.88531e-02,-3.22427e-06,5.47754e-02,8.53322e-02,3.63417e-06,5.47752e-02,9.09948e-02,-4.10833e-06,5.47723e-02,9.74739e-02,3.17086e-06,5.47770e-02,1.03223e-01,-3.22145e-06,5.47698e-02,1.09535e-01,1.07852e-06,5.47770e-02,1.15503e-01,-1.23464e-06,5.47689e-02,1.21558e-01,-1.64282e-06,5.47772e-02,1.27790e-01,6.87798e-07,5.47707e-02,1.33600e-01,-3.68010e-06,5.47757e-02,1.40040e-01,1.62038e-06,5.47734e-02,1.45712e-01,-4.13488e-06,5.47726e-02,1.52213e-01,1.15955e-06,5.47634e-02,1.57947e-01,-2.51007e-06,5.47467e-02,1.64226e-01,-1.50918e-07,5.47312e-02,1.69824e-01,-3.07649e-06,5.46928e-02,1.72615e-01,-1.97350e-05,6.08350e-02,3.68945e-03,-1.47458e-05,6.08072e-02,6.50509e-03,-1.23682e-06,6.08133e-02,1.21028e-02,3.88210e-07,6.08210e-02,1.83909e-02,4.99419e-07,6.08256e-02,2.40796e-02,1.33760e-07,6.08265e-02,3.06536e-02,1.44323e-06,6.08279e-02,3.62466e-02,-6.85311e-07,6.08250e-02,4.27532e-02,1.71931e-06,6.08296e-02,4.85121e-02,-6.39596e-07,6.08244e-02,5.47771e-02,8.32210e-07,6.08301e-02,6.08301e-02,6.04865e-07,6.08246e-02,6.67770e-02,-1.12069e-06,6.08295e-02,7.31370e-02,2.43480e-06,6.08256e-02,7.88238e-02,-3.07306e-06,6.08279e-02,8.53727e-02,3.65726e-06,6.08278e-02,9.09542e-02,-3.96783e-06,6.08259e-02,9.75031e-02,3.27135e-06,6.08295e-02,1.03190e-01,-3.05779e-06,6.08250e-02,1.09550e-01,1.39179e-06,5.17570e-01,6.34024e-01,9.68586e-04,6.08244e-02,1.21550e-01,-1.23466e-06,6.08297e-02,1.27815e-01,1.36366e-06,6.08247e-02,1.33574e-01,-3.19890e-06,6.08281e-02,1.40080e-01,2.45032e-06,6.08263e-02,1.45673e-01,-3.68692e-06,6.08260e-02,1.52247e-01,1.97946e-06,6.08210e-02,1.57936e-01,-1.95794e-06,6.08134e-02,1.64224e-01,1.59660e-07,6.08068e-02,1.69822e-01,-2.96151e-06,6.07713e-02,1.72638e-01,-2.01085e-05,6.69100e-02,3.70983e-03,-1.64310e-05,6.68641e-02,6.50275e-03,-2.17150e-06,6.68342e-02,1.21009e-02,-1.22873e-06,6.68020e-02,1.83799e-02,-2.40842e-06,6.67850e-02,2.41137e-02,-2.56221e-06,6.67831e-02,3.06157e-02,-2.14420e-06,6.67797e-02,3.62863e-02,-2.85185e-06,6.67886e-02,4.27283e-02,-1.23424e-06,6.67773e-02,4.85370e-02,-2.43067e-06,6.67921e-02,5.47698e-02,-1.05211e-06,6.67770e-02,6.08246e-02,-1.12069e-06,6.67916e-02,6.67916e-02,-1.62481e-06,6.67781e-02,7.31049e-02,7.23044e-07,6.67872e-02,7.88520e-02,-2.27267e-06,6.67818e-02,8.53328e-02,2.35166e-06,6.67816e-02,9.09940e-02,-2.09279e-06,6.67872e-02,9.74749e-02,2.95390e-06,6.67779e-02,1.03222e-01,-7.23900e-07,6.67916e-02,1.09535e-01,2.31958e-06,6.67770e-02,1.15502e-01,1.39187e-06,6.67922e-02,1.21557e-01,1.07853e-06,6.67776e-02,1.27790e-01,3.45877e-06,6.67886e-02,1.33599e-01,2.15349e-07,6.67801e-02,1.40041e-01,4.71356e-06,6.67830e-02,1.45711e-01,4.08456e-07,6.67852e-02,1.52213e-01,4.68819e-06,6.68019e-02,1.57947e-01,1.12775e-06,6.68343e-02,1.64226e-01,1.88088e-06,6.68638e-02,1.69824e-01,-2.21225e-06,6.68478e-02,1.72617e-01,-2.04692e-05,7.30007e-02,3.69328e-03,-1.73150e-05,7.29943e-02,6.50067e-03,-8.95060e-07,7.30352e-02,1.21015e-02,2.35379e-06,7.30845e-02,1.83906e-02,3.16916e-06,7.31119e-02,2.40862e-02,4.33472e-06,7.31165e-02,3.06470e-02,3.96729e-06,7.31250e-02,3.62532e-02,3.54851e-06,7.31065e-02,4.27496e-02,3.51116e-06,7.31344e-02,4.85159e-02,2.78417e-06,7.31044e-02,5.47773e-02,2.51990e-06,7.31370e-02,6.08295e-02,2.43532e-06,7.31049e-02,6.67781e-02,7.22948e-07,7.31322e-02,7.31322e-02,2.40231e-06,7.31095e-02,7.88274e-02,-1.29137e-06,7.31216e-02,8.53659e-02,2.16774e-06,7.31215e-02,9.09610e-02,-2.85293e-06,7.31098e-02,9.74997e-02,1.18128e-06,-5.41687e-01,-4.06946e-01,1.05103e-03,7.31051e-02,1.09549e-01,-7.23708e-07,7.31371e-02,1.15498e-01,-3.05832e-06,7.31044e-02,1.21550e-01,-3.22136e-06,7.31345e-02,1.27811e-01,-2.31103e-06,7.31063e-02,1.33578e-01,-5.31924e-06,7.31252e-02,1.40074e-01,-1.89539e-06,7.31164e-02,1.45680e-01,-6.53324e-06,7.31122e-02,1.52241e-01,-2.22673e-06,7.30846e-02,1.57936e-01,-4.84502e-06,7.30353e-02,1.64226e-01,-1.77604e-06,7.29939e-02,1.69826e-01,-3.64580e-06,7.29371e-02,1.72634e-01,-2.07403e-05,7.90669e-02,3.70696e-03,-1.88931e-05,7.90033e-02,6.49898e-03,-3.32439e-06,7.89432e-02,1.21004e-02,-2.79638e-06,7.88751e-02,1.83847e-02,-5.56155e-06,7.88404e-02,2.41049e-02,-5.04880e-06,7.88369e-02,3.06244e-02,-6.09738e-06,7.88312e-02,3.62798e-02,-4.86466e-06,7.88468e-02,4.27361e-02,-4.60393e-06,7.88252e-02,4.85332e-02,-4.29565e-06,7.88531e-02,5.47723e-02,-3.22430e-06,7.88238e-02,6.08256e-02,-3.07322e-06,7.88520e-02,6.67872e-02,-2.27273e-06,7.88274e-02,7.31095e-02,-1.29137e-06,7.88440e-02,7.88440e-02,-1.52339e-06,7.88337e-02,8.53406e-02,6.41686e-07,7.88337e-02,9.09863e-02,-4.65616e-07,7.88440e-02,9.74830e-02,2.17532e-06,7.88273e-02,1.03217e-01,1.18114e-06,7.88521e-02,1.09540e-01,2.95397e-06,7.88239e-02,1.15501e-01,3.27136e-06,7.88531e-02,1.21555e-01,3.17081e-06,7.88254e-02,1.27793e-01,5.34517e-06,7.88469e-02,1.33591e-01,3.39945e-06,7.88314e-02,1.40047e-01,6.84543e-06,7.88369e-02,1.45703e-01,4.12048e-06,7.88405e-02,1.52222e-01,7.35878e-06,7.88751e-02,1.57942e-01,4.12408e-06,7.89433e-02,1.64227e-01,3.48508e-06,7.90030e-02,1.69828e-01,-1.29814e-06,7.90042e-02,1.72620e-01,-2.02578e-05,8.51635e-02,3.70125e-03,-1.91764e-05,8.51687e-02,6.49814e-03,-8.06864e-07,8.52305e-02,1.21011e-02,3.63802e-06,8.53051e-02,1.83889e-02,4.82861e-06,8.53416e-02,2.40965e-02,7.19725e-06,8.53463e-02,3.06341e-02,5.27298e-06,8.53561e-02,3.62677e-02,6.48124e-06,8.53371e-02,4.27450e-02,4.50056e-06,8.53690e-02,4.85241e-02,5.20741e-06,8.53322e-02,5.47754e-02,3.63413e-06,8.53727e-02,6.08279e-02,3.65710e-06,8.53328e-02,6.67818e-02,2.35161e-06,8.53659e-02,7.31216e-02,2.16763e-06,8.53406e-02,7.88337e-02,6.41686e-07,8.53511e-02,8.53511e-02,8.45372e-07,8.53511e-02,9.09760e-02,-1.20744e-06,8.53407e-02,9.74933e-02,-4.65538e-07,8.53659e-02,1.03205e-01,-2.85296e-06,8.53330e-02,1.09545e-01,-2.09277e-06,8.53728e-02,1.15499e-01,-3.96772e-06,8.53322e-02,1.21552e-01,-4.10822e-06,8.53692e-02,1.27803e-01,-4.50183e-06,8.53371e-02,1.33582e-01,-6.13591e-06,8.53563e-02,1.40059e-01,-4.69045e-06,8.53463e-02,1.45693e-01,-7.64119e-06,8.53418e-02,1.52231e-01,-4.98547e-06,8.53051e-02,1.57938e-01,-6.46458e-06,8.52306e-02,1.64226e-01,-2.99232e-06,8.51683e-02,1.69829e-01,-3.85759e-06,8.51003e-02,1.72626e-01,-2.03541e-05,9.12267e-02,3.70093e-03,-2.03530e-05,9.11586e-02,6.49811e-03,-3.86423e-06,9.10964e-02,1.21011e-02,-2.99703e-06,9.10219e-02,1.83889e-02,-6.46102e-06,9.09852e-02,2.40964e-02,-4.98562e-06,9.09807e-02,3.06341e-02,-7.64099e-06,9.09707e-02,3.62677e-02,-4.69018e-06,9.09899e-02,4.27449e-02,-6.13604e-06,9.09578e-02,4.85240e-02,-4.50132e-06,9.09948e-02,5.47752e-02,-4.10818e-06,9.09542e-02,6.08278e-02,-3.96765e-06,9.09940e-02,6.67816e-02,-2.09279e-06,9.09610e-02,7.31215e-02,-2.85282e-06,9.09863e-02,7.88337e-02,-4.65574e-07,9.09759e-02,8.53510e-02,-1.20758e-06,9.09759e-02,9.09759e-02,8.45424e-07,9.09864e-02,9.74932e-02,6.41722e-07,9.09611e-02,1.03205e-01,2.16760e-06,9.09942e-02,1.09545e-01,2.35163e-06,9.09543e-02,1.15499e-01,3.65729e-06,9.09948e-02,1.21552e-01,3.63410e-06,9.09579e-02,1.27803e-01,5.20731e-06,9.09898e-02,1.33582e-01,4.50066e-06,9.09709e-02,1.40059e-01,6.48141e-06,9.09807e-02,1.45693e-01,5.27391e-06,9.09854e-02,1.52231e-01,7.19634e-06,9.10219e-02,1.57938e-01,4.82485e-06,9.10965e-02,1.64226e-01,3.63802e-06,9.11583e-02,1.69829e-01,-8.13640e-07,9.11634e-02,1.72626e-01,-1.91749e-05,9.73228e-02,3.70721e-03,-2.02581e-05,9.73240e-02,6.49900e-03,-1.29869e-06,9.73837e-02,1.21004e-02,3.48460e-06,9.74519e-02,1.83848e-02,4.12463e-06,9.74865e-02,2.41050e-02,7.35830e-06,9.74901e-02,3.06244e-02,4.12035e-06,9.74955e-02,3.62799e-02,6.84518e-06,9.74801e-02,4.27361e-02,3.39940e-06,9.75016e-02,4.85335e-02,5.34533e-06,9.74739e-02,5.47723e-02,3.17093e-06,9.75031e-02,6.08259e-02,3.27141e-06,9.74749e-02,6.67872e-02,2.95393e-06,9.74997e-02,7.31098e-02,1.18128e-06,9.74830e-02,7.88440e-02,2.17536e-06,9.74933e-02,8.53407e-02,-4.65485e-07,9.74932e-02,9.09864e-02,6.41675e-07,9.74830e-02,9.74830e-02,-1.52334e-06,9.74996e-02,1.03217e-01,-1.29126e-06,9.74750e-02,1.09540e-01,-2.27269e-06,9.75032e-02,1.15501e-01,-3.07313e-06,9.74739e-02,1.21555e-01,-3.22427e-06,9.75018e-02,1.27794e-01,-4.29529e-06,9.74802e-02,1.33591e-01,-4.60397e-06,9.74958e-02,1.40047e-01,-4.86494e-06,9.74901e-02,1.45703e-01,-6.09751e-06,9.74866e-02,1.52222e-01,-5.04880e-06,9.74519e-02,1.57942e-01,-5.56059e-06,9.73838e-02,1.64227e-01,-2.79486e-06,9.73237e-02,1.69828e-01,-3.32117e-06,9.72601e-02,1.72620e-01,-1.88929e-05,1.03390e-01,3.69302e-03,-2.07402e-05,1.03333e-01,6.50063e-03,-3.64491e-06,1.03292e-01,1.21014e-02,-1.77706e-06,1.03242e-01,1.83906e-02,-4.84599e-06,1.03215e-01,2.40861e-02,-2.22727e-06,1.03211e-01,3.06469e-02,-6.53390e-06,1.03202e-01,3.62532e-02,-1.89542e-06,1.03221e-01,4.27494e-02,-5.31916e-06,1.03193e-01,4.85159e-02,-2.31074e-06,1.03223e-01,5.47770e-02,-3.22161e-06,1.03190e-01,6.08295e-02,-3.05824e-06,1.03222e-01,6.67779e-02,-7.23837e-07,-5.22070e+00,-2.91202e+00,-2.36389e-02,1.03217e-01,7.88273e-02,1.18109e-06,1.03205e-01,8.53659e-02,-2.85282e-06,1.03205e-01,9.09611e-02,2.16753e-06,1.03217e-01,9.74996e-02,-1.29137e-06,1.03195e-01,1.03195e-01,2.40196e-06,1.03222e-01,1.09549e-01,7.23166e-07,1.03190e-01,1.15498e-01,2.43522e-06,1.03223e-01,1.21550e-01,2.51986e-06,1.03193e-01,1.27811e-01,2.78384e-06,1.03220e-01,1.33577e-01,3.51114e-06,1.03202e-01,1.40074e-01,3.54789e-06,1.03210e-01,1.45680e-01,3.96709e-06,1.03215e-01,1.52241e-01,4.33485e-06,1.03242e-01,1.57936e-01,3.17018e-06,1.03292e-01,1.64226e-01,2.35439e-06,1.03333e-01,1.69826e-01,-8.95339e-07,1.03326e-01,1.72634e-01,-1.73152e-05,1.09479e-01,3.70991e-03,-2.04692e-05,1.09463e-01,6.50275e-03,-2.21231e-06,1.09493e-01,1.21009e-02,1.88096e-06,1.09525e-01,1.83800e-02,1.12790e-06,1.09542e-01,2.41137e-02,4.68859e-06,1.09544e-01,3.06158e-02,4.08155e-07,1.09547e-01,3.62865e-02,4.71413e-06,1.09538e-01,4.27283e-02,2.15325e-07,1.09549e-01,4.85373e-02,3.45879e-06,1.09535e-01,5.47698e-02,1.07857e-06,1.09550e-01,6.08250e-02,1.39179e-06,1.09535e-01,6.67916e-02,2.31959e-06,1.09549e-01,7.31051e-02,-7.23900e-07,1.09540e-01,7.88521e-02,2.95392e-06,1.09545e-01,8.53330e-02,-2.09276e-06,1.09545e-01,9.09942e-02,2.35168e-06,1.09540e-01,9.74750e-02,-2.27268e-06,1.09549e-01,1.03222e-01,7.23128e-07,1.09535e-01,1.09535e-01,-1.62484e-06,1.09550e-01,1.15502e-01,-1.12079e-06,1.09535e-01,1.21557e-01,-1.05200e-06,1.09550e-01,1.27790e-01,-2.43051e-06,1.09538e-01,1.33599e-01,-1.23424e-06,1.09547e-01,1.40041e-01,-2.85191e-06,1.09544e-01,1.45711e-01,-2.14406e-06,1.09542e-01,1.52213e-01,-2.56205e-06,1.09525e-01,1.57947e-01,-2.40809e-06,1.09493e-01,1.64226e-01,-1.22791e-06,1.09463e-01,1.69824e-01,-2.17156e-06,1.09417e-01,1.72617e-01,-1.64310e-05,1.15556e-01,3.68936e-03,-2.01085e-05,1.15520e-01,6.50507e-03,-2.96140e-06,1.15514e-01,1.21028e-02,1.59755e-07,1.15506e-01,1.83909e-02,-1.95772e-06,1.15501e-01,2.40796e-02,1.97969e-06,1.15501e-01,3.06536e-02,-3.68725e-06,1.15499e-01,3.62466e-02,2.45012e-06,1.15502e-01,4.27531e-02,-3.19880e-06,1.15497e-01,4.85121e-02,1.36297e-06,1.15503e-01,5.47770e-02,-1.23470e-06,1.85825e+00,2.63335e+00,-1.30221e-02,1.15502e-01,6.67770e-02,1.39187e-06,1.15498e-01,7.31371e-02,-3.05850e-06,1.15501e-01,7.88239e-02,3.27125e-06,1.15499e-01,8.53728e-02,-3.96803e-06,1.15499e-01,9.09543e-02,3.65730e-06,1.15501e-01,9.75032e-02,-3.07316e-06,1.15498e-01,1.03190e-01,2.43478e-06,1.15502e-01,1.09550e-01,-1.12079e-06,1.15497e-01,1.15497e-01,6.05101e-07,1.15503e-01,1.21550e-01,8.32094e-07,1.15497e-01,1.27815e-01,-6.39596e-07,1.15502e-01,1.33574e-01,1.71935e-06,1.15499e-01,1.40080e-01,-6.85404e-07,1.15501e-01,1.45673e-01,1.44309e-06,1.15501e-01,1.52247e-01,1.33555e-07,1.15506e-01,1.57936e-01,4.99118e-07,1.15514e-01,1.64224e-01,3.87985e-07,1.15520e-01,1.69822e-01,-1.23676e-06,1.15492e-01,1.72638e-01,-1.47458e-05,1.21634e-01,3.71203e-03,-1.97350e-05,1.21596e-01,6.50257e-03,-3.07651e-06,1.21580e-01,1.21011e-02,-1.51099e-07,1.21564e-01,1.83795e-02,-2.51029e-06,1.21554e-01,2.41141e-02,1.15955e-06,1.21554e-01,3.06149e-02,-4.13504e-06,1.21551e-01,3.62868e-02,1.62044e-06,1.21556e-01,4.27268e-02,-3.68017e-06,1.21550e-01,4.85373e-02,6.87982e-07,1.21558e-01,5.47689e-02,-1.64282e-06,1.21550e-01,6.08244e-02,-1.23475e-06,1.21557e-01,6.67922e-02,1.07853e-06,1.21550e-01,7.31044e-02,-3.22148e-06,1.21555e-01,7.88531e-02,3.17085e-06,1.21552e-01,8.53322e-02,-4.10822e-06,1.21552e-01,9.09948e-02,3.63410e-06,1.21555e-01,9.74739e-02,-3.22427e-06,1.21550e-01,1.03223e-01,2.51987e-06,1.21557e-01,1.09535e-01,-1.05200e-06,1.21550e-01,1.15503e-01,8.32199e-07,1.21558e-01,1.21558e-01,1.14985e-06,1.21550e-01,1.27790e-01,-3.26538e-07,1.21556e-01,1.33600e-01,2.33087e-06,1.21552e-01,1.40040e-01,-3.92715e-07,1.21553e-01,1.45712e-01,2.23565e-06,1.21555e-01,1.52213e-01,4.51288e-07,1.21563e-01,1.57947e-01,1.19241e-06,1.21580e-01,1.64226e-01,6.58776e-07,1.21596e-01,1.69824e-01,-8.44263e-07,1.21572e-01,1.72615e-01,-1.31721e-05,1.27722e-01,3.69203e-03,-1.86050e-05,1.27709e-01,6.50215e-03,-2.21511e-06,1.27739e-01,1.21017e-02,1.83345e-06,1.27776e-01,1.83905e-02,6.09029e-07,1.27797e-01,2.40830e-02,5.58831e-06,1.27800e-01,3.06499e-02,-9.31021e-07,1.27806e-01,3.62500e-02,6.25411e-06,1.27792e-01,4.27504e-02,-1.09276e-06,1.05140e-01,1.15903e-03,3.95212e-04,1.27790e-01,5.47772e-02,6.88011e-07,1.27815e-01,6.08297e-02,1.36302e-06,1.27790e-01,6.67776e-02,3.45903e-06,1.27811e-01,7.31345e-02,-2.31103e-06,1.27793e-01,7.88254e-02,5.34483e-06,1.27803e-01,8.53692e-02,-4.50205e-06,1.27803e-01,9.09579e-02,5.20698e-06,1.27794e-01,9.75018e-02,-4.29563e-06,1.27811e-01,1.03193e-01,2.78422e-06,1.27790e-01,1.09550e-01,-2.43034e-06,1.27815e-01,1.15497e-01,-6.39596e-07,1.27790e-01,1.21550e-01,-3.26388e-07,1.27813e-01,1.27813e-01,-3.31010e-06,1.27792e-01,1.33577e-01,4.39108e-07,1.27806e-01,1.40077e-01,-4.15691e-06,1.27800e-01,1.45677e-01,-4.82982e-07,1.27797e-01,1.52244e-01,-3.41172e-06,1.27776e-01,1.57937e-01,-1.62238e-06,1.27739e-01,1.64225e-01,-1.25756e-06,1.27709e-01,1.69825e-01,-1.35920e-06,1.27658e-01,1.72635e-01,-1.14864e-05,1.33790e-01,3.70997e-03,-1.79894e-05,1.33732e-01,6.50054e-03,-3.50349e-06,1.33679e-01,1.21003e-02,-1.41285e-06,1.33620e-01,1.83824e-02,-4.85905e-06,1.33589e-01,2.41078e-02,-9.82035e-07,1.33586e-01,3.06204e-02,-7.24785e-06,1.33581e-01,3.62824e-02,-2.43102e-07,1.33595e-01,4.27320e-02,-6.51668e-06,1.33577e-01,4.85351e-02,-1.09294e-06,1.33600e-01,5.47707e-02,-3.68015e-06,1.33574e-01,6.08247e-02,-3.19890e-06,1.33599e-01,6.67886e-02,2.15416e-07,1.33578e-01,7.31063e-02,-5.31911e-06,1.33591e-01,7.88469e-02,3.39937e-06,1.33582e-01,8.53371e-02,-6.13589e-06,1.33582e-01,9.09898e-02,4.50066e-06,1.33591e-01,9.74802e-02,-4.60392e-06,1.33577e-01,1.03220e-01,3.51099e-06,1.33599e-01,1.09538e-01,-1.23426e-06,1.33574e-01,1.15502e-01,1.71963e-06,1.33600e-01,1.21556e-01,2.33088e-06,1.33577e-01,1.27792e-01,4.39026e-07,1.33595e-01,1.33595e-01,4.56368e-06,1.33581e-01,1.40045e-01,5.43295e-07,1.33586e-01,1.45707e-01,4.99483e-06,1.33589e-01,1.52219e-01,1.87054e-06,1.33620e-01,1.57944e-01,3.42094e-06,1.33679e-01,1.64227e-01,1.71611e-06,1.33731e-01,1.69826e-01,2.32936e-07,1.33728e-01,1.72617e-01,-9.34249e-06,1.39885e-01,3.70148e-03,-1.63683e-05,1.39889e-01,6.49901e-03,-1.79229e-06,1.39945e-01,1.21013e-02,2.49556e-06,1.40015e-01,1.83890e-02,1.66788e-06,1.40050e-01,2.40925e-02,7.04437e-06,1.40054e-01,3.06374e-02,1.83953e-07,4.33471e-01,3.79988e-01,-9.08834e-04,1.40045e-01,4.27464e-02,-2.43095e-07,1.40077e-01,4.85208e-02,6.25444e-06,1.40040e-01,5.47757e-02,1.62050e-06,1.40080e-01,6.08281e-02,2.45057e-06,1.40041e-01,6.67801e-02,4.71356e-06,1.40074e-01,7.31252e-02,-1.89539e-06,1.40047e-01,7.88314e-02,6.84524e-06,1.40059e-01,8.53563e-02,-4.69087e-06,1.40059e-01,9.09709e-02,6.48177e-06,1.40047e-01,9.74958e-02,-4.86494e-06,1.40074e-01,1.03202e-01,3.54802e-06,1.40041e-01,1.09547e-01,-2.85175e-06,1.40080e-01,1.15499e-01,-6.85404e-07,1.40040e-01,1.21552e-01,-3.92715e-07,1.40077e-01,1.27806e-01,-4.15698e-06,1.40045e-01,1.33581e-01,5.43391e-07,1.40065e-01,1.40065e-01,-5.50217e-06,1.40054e-01,1.45690e-01,-6.09313e-07,1.40050e-01,1.52235e-01,-4.88564e-06,1.40015e-01,1.57938e-01,-2.14423e-06,1.39945e-01,1.64226e-01,-1.91188e-06,1.39888e-01,1.69828e-01,-9.96214e-07,1.39823e-01,1.72626e-01,-7.59291e-06,1.45951e-01,3.81608e-03,-1.58601e-05,1.45884e-01,6.51322e-03,-3.33803e-06,1.45816e-01,1.21005e-02,-1.36772e-06,1.45739e-01,1.83869e-02,-4.95457e-06,1.45702e-01,2.40991e-02,-6.63250e-07,1.45698e-01,3.06290e-02,-7.72851e-06,1.45690e-01,3.62725e-02,1.83956e-07,1.45707e-01,4.27411e-02,-7.24799e-06,1.45677e-01,4.85266e-02,-9.31122e-07,1.45712e-01,5.47734e-02,-4.13494e-06,1.45673e-01,6.08263e-02,-3.68684e-06,1.45711e-01,6.67830e-02,4.08456e-07,1.45680e-01,7.31164e-02,-6.53261e-06,1.45703e-01,7.88369e-02,4.12016e-06,1.45693e-01,8.53463e-02,-7.64102e-06,1.45693e-01,9.09807e-02,5.27309e-06,1.45703e-01,9.74901e-02,-6.09745e-06,1.45680e-01,1.03210e-01,3.96701e-06,1.45711e-01,1.09544e-01,-2.14409e-06,1.45673e-01,1.15501e-01,1.44379e-06,1.45712e-01,1.21553e-01,2.23558e-06,1.45677e-01,1.27800e-01,-4.83182e-07,1.45707e-01,1.33586e-01,4.99483e-06,1.45690e-01,1.40054e-01,-6.09362e-07,1.45698e-01,1.45698e-01,5.41240e-06,1.45702e-01,1.52228e-01,9.23473e-07,1.45739e-01,1.57940e-01,3.55335e-06,1.45816e-01,1.64226e-01,1.47997e-06,1.45884e-01,1.69814e-01,7.89541e-07,1.45887e-01,1.72511e-01,-5.20311e-06,1.52044e-01,4.03141e-03,-1.51352e-05,1.52036e-01,6.58380e-03,-2.47010e-06,1.52086e-01,1.21242e-02,1.43969e-06,1.52173e-01,1.83743e-02,8.59380e-07,1.52221e-01,2.41056e-02,6.10018e-06,1.52228e-01,3.06252e-02,-6.63222e-07,1.52235e-01,3.62770e-02,7.04443e-06,1.52219e-01,4.27382e-02,-9.82156e-07,1.52244e-01,4.85303e-02,5.58845e-06,1.52213e-01,5.47726e-02,1.15955e-06,1.52247e-01,6.08260e-02,1.97946e-06,1.52213e-01,6.67852e-02,4.68840e-06,1.52241e-01,7.31122e-02,-2.22797e-06,1.52222e-01,7.88405e-02,7.35847e-06,1.52231e-01,8.53418e-02,-4.98522e-06,1.52231e-01,9.09854e-02,7.19634e-06,1.52222e-01,9.74866e-02,-5.04888e-06,1.52241e-01,1.03215e-01,4.33483e-06,1.52213e-01,1.09542e-01,-2.56205e-06,1.52247e-01,1.15501e-01,1.33794e-07,1.52213e-01,1.21555e-01,4.51288e-07,1.52244e-01,1.27797e-01,-3.41172e-06,1.52219e-01,1.33589e-01,1.87050e-06,1.52235e-01,1.40050e-01,-4.88570e-06,1.52228e-01,1.45702e-01,9.23473e-07,1.52221e-01,1.52221e-01,-4.27217e-06,1.52173e-01,1.57953e-01,-9.11507e-07,1.52086e-01,1.64203e-01,-1.09777e-06,1.52035e-01,1.69743e-01,1.45504e-07,1.51977e-01,1.72296e-01,-2.95306e-06,1.58112e-01,4.42089e-03,-1.53721e-05,1.58072e-01,6.76313e-03,-3.83536e-06,1.58044e-01,1.21551e-02,-4.27264e-07,1.57993e-01,1.83339e-02,-2.14183e-06,1.57953e-01,2.41540e-02,8.59370e-07,1.57940e-01,3.05875e-02,-4.95455e-06,1.57938e-01,3.63123e-02,1.66786e-06,1.57945e-01,4.27073e-02,-4.85907e-06,1.57937e-01,4.85512e-02,6.08986e-07,1.57947e-01,5.47634e-02,-2.51015e-06,1.57936e-01,6.08210e-02,-1.95776e-06,1.57947e-01,6.68019e-02,1.12776e-06,1.57936e-01,7.30846e-02,-4.84565e-06,1.57942e-01,7.88751e-02,4.12590e-06,1.57938e-01,8.53051e-02,-6.46458e-06,1.57938e-01,9.10219e-02,4.82485e-06,1.57942e-01,9.74519e-02,-5.56059e-06,1.57936e-01,1.03242e-01,3.16942e-06,1.57947e-01,1.09525e-01,-2.40822e-06,1.57936e-01,1.15506e-01,4.99347e-07,1.57947e-01,1.21563e-01,1.19238e-06,1.57937e-01,1.27776e-01,-1.62238e-06,1.57944e-01,1.33620e-01,3.42101e-06,1.57938e-01,1.40015e-01,-2.14425e-06,1.57940e-01,1.45739e-01,3.55331e-06,1.57953e-01,1.52173e-01,-9.11496e-07,1.57993e-01,1.57993e-01,1.49982e-06,1.58044e-01,1.64172e-01,3.48708e-07,1.58071e-01,1.69564e-01,1.38884e-06,1.58038e-01,1.71906e-01,2.88366e-07,1.64022e-01,4.87923e-03,-1.69262e-05,1.64055e-01,6.99782e-03,-5.45667e-06,1.64124e-01,1.22033e-02,-6.76566e-07,1.64172e-01,1.82831e-02,-4.27269e-07,1.64203e-01,2.42411e-02,1.43968e-06,1.64227e-01,3.05113e-02,-1.36772e-06,1.64226e-01,3.63816e-02,2.49552e-06,1.64227e-01,4.26485e-02,-1.41289e-06,1.64225e-01,4.85879e-02,1.83346e-06,1.64226e-01,5.47467e-02,-1.51041e-07,1.64224e-01,6.08134e-02,1.59816e-07,1.64226e-01,6.68343e-02,1.88062e-06,1.64226e-01,7.30353e-02,-1.77669e-06,1.64227e-01,7.89433e-02,3.48411e-06,1.64226e-01,8.52306e-02,-2.99680e-06,1.64226e-01,9.10965e-02,3.63477e-06,1.64227e-01,9.73838e-02,-2.79394e-06,1.64226e-01,1.03292e-01,2.35343e-06,1.64226e-01,1.09493e-01,-1.22791e-06,1.64224e-01,1.15514e-01,3.88184e-07,1.64226e-01,1.21580e-01,6.58861e-07,1.64225e-01,1.27739e-01,-1.25756e-06,1.64227e-01,1.33679e-01,1.71611e-06,1.64226e-01,1.39945e-01,-1.91188e-06,1.64226e-01,1.45816e-01,1.47996e-06,1.64203e-01,1.52086e-01,-1.09777e-06,1.64172e-01,1.58044e-01,3.48705e-07,1.64124e-01,1.64124e-01,4.73749e-07,1.64054e-01,1.69329e-01,3.09525e-06,1.63951e-01,1.71443e-01,5.31521e-06,1.69008e-01,5.00525e-03,-2.12157e-05,1.69088e-01,7.23910e-03,-1.10463e-05,1.69329e-01,1.22721e-02,-5.45667e-06,1.69564e-01,1.82550e-02,-3.83536e-06,1.69743e-01,2.42914e-02,-2.47009e-06,1.69814e-01,3.04431e-02,-3.33803e-06,1.69828e-01,3.64384e-02,-1.79230e-06,1.69826e-01,4.25954e-02,-3.50349e-06,1.69825e-01,4.86180e-02,-2.21511e-06,1.69824e-01,5.47312e-02,-3.07648e-06,1.69822e-01,6.08068e-02,-2.96151e-06,1.69824e-01,6.68638e-02,-2.21211e-06,1.69826e-01,7.29939e-02,-3.64469e-06,1.69828e-01,7.90030e-02,-1.29776e-06,1.69829e-01,8.51683e-02,-3.85873e-06,1.69829e-01,9.11583e-02,-8.05151e-07,1.69828e-01,9.73237e-02,-3.32042e-06,1.69826e-01,1.03333e-01,-8.95997e-07,1.69824e-01,1.09463e-01,-2.17152e-06,1.69822e-01,1.15520e-01,-1.23676e-06,1.69824e-01,1.21596e-01,-8.44343e-07,1.69825e-01,1.27709e-01,-1.35923e-06,1.69826e-01,1.33731e-01,2.32944e-07,1.69828e-01,1.39888e-01,-9.96213e-07,1.69814e-01,1.45884e-01,7.89537e-07,1.69743e-01,1.52035e-01,1.45502e-07,1.69564e-01,1.58071e-01,1.38884e-06,1.69329e-01,1.64054e-01,3.09525e-06,1.69088e-01,1.69088e-01,7.73485e-06,1.68976e-01,1.71309e-01,1.27455e-05,1.71632e-01,4.69451e-03,-2.91226e-05,1.71322e-01,7.31861e-03,-2.12157e-05,1.71448e-01,1.23051e-02,-1.69262e-05,1.71906e-01,1.82148e-02,-1.53721e-05,1.72296e-01,2.42825e-02,-1.51352e-05,1.72511e-01,3.03759e-02,-1.58601e-05,1.72626e-01,3.64417e-02,-1.63683e-05,1.72617e-01,4.25369e-02,-1.79894e-05,1.72635e-01,4.86054e-02,-1.86050e-05,1.72615e-01,5.46928e-02,-1.97350e-05,1.72638e-01,6.07713e-02,-2.01085e-05,1.72617e-01,6.68478e-02,-2.04692e-05,1.72634e-01,7.29371e-02,-2.07403e-05,1.72620e-01,7.90042e-02,-2.02582e-05,1.72626e-01,8.51003e-02,-2.03580e-05,1.72626e-01,9.11634e-02,-1.91749e-05,1.72620e-01,9.72601e-02,-1.88935e-05,1.72634e-01,1.03326e-01,-1.73150e-05,1.72617e-01,1.09417e-01,-1.64310e-05,1.72638e-01,1.15492e-01,-1.47458e-05,1.72615e-01,1.21572e-01,-1.31721e-05,1.72635e-01,1.27658e-01,-1.14864e-05,1.72617e-01,1.33728e-01,-9.34249e-06,1.72626e-01,1.39823e-01,-7.59291e-06,1.72511e-01,1.45887e-01,-5.20311e-06,1.72296e-01,1.51977e-01,-2.95306e-06,1.71906e-01,1.58038e-01,2.88366e-07,1.71443e-01,1.63951e-01,5.31521e-06,1.71309e-01,1.68976e-01,1.27455e-05,1.71632e-01,1.71632e-01,2.03921e-05,])
-
-    plot_quantity(xs0, us, r"$u_s$")
-    plot_quantity(xa0, ua, r"$u_a$")
-    plot_quantity(xa0, fa, r"$f_a$")
-    plot_quantity(xs0, fs, r"$f_s$")
-
-    ua_diff = ua_gpu - ua
-    abs_err = np.linalg.norm(ua_gpu - ua)
-    rel_err = abs_err / np.linalg.norm(ua)
-    print(f"{ua_diff=}")
-    ind = np.argmax(np.abs(ua_diff))
-    print(f"{ind=} {ua_gpu[ind]=} {ua[ind]=}")
-    print(f"{abs_err=}")
-    print(f"{rel_err=}")
